refactor: Move gmae.vue imported components to a GameSystem folder
