+++ src/components/GameSystem/EffectWindow.vue
