+++ src/components/GameSystem/hintbutton.vue
